refactor: `findMainPkg` fn
